code refactor into different package, agent for LLM agent pattern, orchestrator for orchestrator pattern and spell for the previous project
