+++ src/main/agent/LLMAgent.java
@@ -1,4 +1,4 @@
-package main;
+package main.agent;
 
 public interface LLMAgent {
     String interact(String input);
