+++ src/main/orchestrator/PromptOrchestrator.java
@@ -1,4 +1,4 @@
-package main;
+package main.orchestrator;
 
 public interface PromptOrchestrator {
     String orchestrate(String agentType, String input);
