+++ src/main/spell/CastSpellCommand.java
@@ -1,11 +1,11 @@
-package main;
+package main.spell;
 
 public class CastSpellCommand implements Command {
 
     private Spell spell;
 
     /**
-     * Constructor for main.CastSpellCommand
+     * Constructor for main.spell.CastSpellCommand
      * @param spell
      */
     public CastSpellCommand(Spell spell) {
@@ -25,6 +25,6 @@ public class CastSpellCommand implements Command {
     public void undo() {
         System.out.println("Undoing spell... ("+ spell.getClass().getSimpleName()+ ")");
         spell.setState(new ChargedState());
-        System.out.println("main.Spell is now inactive!");
+        System.out.println("main.spell.Spell is now inactive!");
     }
 }
