+++ src/main/spell/ChargedState.java
@@ -1,4 +1,4 @@
-package main;
+package main.spell;
 
 public class ChargedState implements SpellState {
     /**
