+++ src/main/spell/Command.java
@@ -1,4 +1,4 @@
-package main;
+package main.spell;
 
 public interface Command {
     /**
