+++ src/main/spell/FireBallSpell.java
@@ -1,4 +1,4 @@
-package main;
+package main.spell;
 
 import java.util.Timer;
 import java.util.TimerTask;
