+++ src/main/spell/Spell.java
@@ -1,4 +1,4 @@
-package main;
+package main.spell;
 
 public interface Spell {
     void cast();
