+++ src/main/spell/SpellLogger.java
@@ -1,0 +1,8 @@
+package main.spell;
+
+// Observer Pattern: Observes changes in main.spell.SpellBook and logs them
+public class SpellLogger implements SpellStatusObserver {
+    public void onSpellStatusChange(Spell spell, String status) {
+        System.out.println("main.spell.Spell status changed: " + spell.getClass().getSimpleName() + " : " + status);
+    }
+}
