+++ src/main/spell/SpellState.java
@@ -1,4 +1,4 @@
-package main;
+package main.spell;
 
 public interface SpellState {
     void cast();
