+++ src/main/spell/SpellStatusObserver.java
@@ -1,4 +1,4 @@
-package main;
+package main.spell;
 
 interface SpellStatusObserver {
     /**
