+++ src/main/spell/TimedSpell.java
@@ -1,4 +1,4 @@
-package main;
+package main.spell;
 
 public interface TimedSpell {
     void startTimer();
