+++ src/main/spell/Wizard.java
@@ -1,4 +1,4 @@
-package main;
+package main.spell;
 
 import java.util.ArrayList;
 import java.util.List;
